Enhance map functionality and use a changelog
Implement a changelog mechanism which saves any modification done
on the map object as an entry appended to the changelog.
Add compactification to the changelog once an arbitrary condition
is met.
Give up on keeping the map entries also in the OMAP structure attached
to the object as the direct changelong implementation is faster.
Add functionality to time the execution of arbitrary pieces of code
in order to evaluate the performance of different implementation
techniques.
Add a series of simple tests cases.